requirements.txt
fastapi
uvicorn
tavily-python
python-dotenv
pydantic